PYIC-8697: Move Environment enum to constants.ts
